fix(shared): move Bedrock embedding into shared BedrockClientManager
Adds generateEmbedding() to BedrockClientManager in @secondlayer/shared
so mcp_backend doesn't need a direct @aws-sdk/client-bedrock-runtime dep.
Removes the dep from mcp_backend/package.json — fixes CI build failure
where the package wasn't resolvable in isolated npm install.
